:sparkles: :: View + if statement
